backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.1.0
celery==5.3.4
redis==4.6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-environ==0.12.0
flake8==6.1.0
black==23.11.0
numpy==2.2.6
tensorflow==2.20.0